pyproject.toml
[project]
name = "vision-agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "vision-agents[gemini,getstream,ultralytics]>=0.3.6",
]
